Refactor: only expect `randomUUID` implementation
Before, the whole `UUID` class was an expected class, which lead to duplication in its actual implementation, and furthermore to compilation problems on Kotlin 1.8.0-RC.
